feat: change long type to int in endpoints/model (#126)
Closes: FHIQC-49

+++ src/main/java/org/folio/holdingsiq/model/PackageCreated.java
@@ -4,4 +4,4 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record PackageCreated(@JsonProperty("packageId") Long packageId) { }
+public record PackageCreated(@JsonProperty("packageId") Integer packageId) { }
