+++ src/main/java/org/folio/holdingsiq/service/TitlesHoldingsIQService.java
@@ -11,13 +11,13 @@ import org.folio.holdingsiq.model.Titles;
 
 public interface TitlesHoldingsIQService {
 
-  CompletableFuture<Title> retrieveTitle(long titleId);
+  CompletableFuture<Title> retrieveTitle(int titleId);
 
   CompletableFuture<Titles> retrieveTitles(String rmapiQuery);
 
   CompletableFuture<Titles> retrieveTitles(FilterQuery filterQuery, String searchType, Sort sort, int page, int count);
 
-  CompletableFuture<Titles> retrieveTitles(long providerId, long packageId, FilterQuery filterQuery, String searchType,
+  CompletableFuture<Titles> retrieveTitles(int providerId, int packageId, FilterQuery filterQuery, String searchType,
                                            Sort sort, int page, int count);
 
   CompletableFuture<Title> postTitle(TitlePost titlePost, PackageId packageId);
